refactor: split data converter format processing into dedicated codec packages with shared normalization utilities
